ch04-observation/pyproject.toml
[project]
name = "chainlit-observation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.4"
dependencies = [
    "chainlit==2.9.6",
    "langchain>=0.1.0",
    "langchain-community>=0.0.20",
    "langchain-openai>=0.0.5",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]
